# frontend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Semenq API specification - AI-powered medicine accessibility platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List user prescriptions
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"

    post:
      operationId: uploadPrescription
      tags: [prescriptions]
      summary: Upload and parse prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Uploaded prescription with parsed medicines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Prescription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
...
components:
  schemas:
...
    Prescription:
      type: object
      required: [id, status, uploadedAt, medicines]
      properties:
        id: { type: integer }
        doctorName: { type: ["string", "null"] }
        patientName: { type: ["string", "null"] }
        hospitalName: { type: ["string", "null"] }
        uploadedAt: { type: string }
        status: { type: string, enum: [processing, parsed, failed] }
        imageUrl: { type: ["string", "null"] }
        overallConfidence: { type: number }
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/ParsedMedicine"
        notes: { type: ["string", "null"] }
    ParsedMedicine:
      type: object
      required: [id, name, confidence]
      properties:
        id: { type: integer }
        name: { type: string }
        dosage: { type: ["string", "null"] }
        frequency: { type: ["string", "null"] }
        duration: { type: ["string", "null"] }
        confidence: { type: number }
        matchedMedicineId: { type: ["integer", "null"] }
        status: { type: string, enum: [confirmed, lowConfidence, unmatched] }
    PrescriptionInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64: { type: string }
        mimeType: { type: string }
...